# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 2.0.0
  description: PharmaPOS Pakistan - Complete Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerPharmacy
      tags: [auth]
      summary: Register a new pharmacy + owner account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Reset password using registered phone verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation or phone mismatch
        "404":
          description: Username not found
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    RegisterInput:
      type: object
      properties:
        pharmacyName:
          type: string
        ownerName:
          type: string
        username:
          type: string
        password:
          type: string
        phone:
          type: string
        city:
          type: string
        address:
          type: string
      required: [pharmacyName, ownerName, username, password, phone]
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    ForgotPasswordInput:
      type: object
      properties:
        username:
          type: string
        phone:
          type: string
        newPassword:
          type: string
          minLength: 6
      required: [username, phone, newPassword]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserPublic"
        pharmacy:
          $ref: "#/components/schemas/Pharmacy"
      required: [user, pharmacy]
    Pharmacy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        ownerName:
          type: string
        licenseNumber:
          type: string
        email:
          type: string
        city:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, isActive, createdAt]
...
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        pharmacyId:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, pharmacyId, username, fullName, role, isActive, createdAt]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
...